feat(3601): redesign vote status & refine LP mechanic to update market
