--- apps/governance/src/routes/proposals/components/proposal-votes-table/index.tsx
@@ -1,1 +1,0 @@
-export { ProposalVotesTable } from './proposal-votes-table';
