--- apps/governance/src/routes/proposals/components/proposal-voting-status/index.tsx
@@ -1,1 +1,0 @@
-export * from './proposal-voting-status';
