+++ apps/governance/src/routes/proposals/components/proposal/proposal.spec.tsx
@@ -33,12 +33,6 @@ jest.mock('../proposal-change-table', () => ({
 jest.mock('../proposal-json', () => ({
   ProposalJson: () => <div data-testid="proposal-json"></div>,
 }));
-jest.mock('../proposal-votes-table', () => ({
-  ProposalVotesTable: () => <div data-testid="proposal-votes-table"></div>,
-}));
-jest.mock('../vote-details', () => ({
-  VoteDetails: () => <div data-testid="proposal-vote-details"></div>,
-}));
 jest.mock('../list-asset', () => ({
   ListAsset: () => <div data-testid="proposal-list-asset"></div>,
 }));
@@ -91,8 +85,6 @@ it('renders each section', async () => {
   expect(await screen.findByTestId('proposal-header')).toBeInTheDocument();
   expect(screen.getByTestId('proposal-change-table')).toBeInTheDocument();
   expect(screen.getByTestId('proposal-json')).toBeInTheDocument();
-  expect(screen.getByTestId('proposal-votes-table')).toBeInTheDocument();
-  expect(screen.getByTestId('proposal-vote-details')).toBeInTheDocument();
   expect(screen.queryByTestId('proposal-list-asset')).not.toBeInTheDocument();
 });
