+++ apps/governance/src/routes/proposals/components/proposals-list-item/proposals-list-item.tsx
@@ -4,28 +4,19 @@ import { ProposalsListItemDetails } from './proposals-list-item-details';
 import { useUserVote } from '../vote-details/use-user-vote';
 import type { ProposalFieldsFragment } from '../../proposals/__generated__/Proposals';
 import type { ProposalQuery } from '../../proposal/__generated__/Proposal';
-import type { NetworkParamsResult } from '@vegaprotocol/network-parameters';
 
 interface ProposalsListItemProps {
   proposal?: ProposalFieldsFragment | ProposalQuery['proposal'] | null;
-  networkParams: Partial<NetworkParamsResult> | null;
 }
 
-export const ProposalsListItem = ({
-  proposal,
-  networkParams,
-}: ProposalsListItemProps) => {
+export const ProposalsListItem = ({ proposal }: ProposalsListItemProps) => {
   const { voteState } = useUserVote(proposal?.id);
-  if (!proposal || !proposal.id || !networkParams) return null;
+  if (!proposal || !proposal.id) return null;
 
   return (
     <li id={proposal.id} data-testid="proposals-list-item">
       <RoundedWrapper paddingBottom={true} heightFull={true}>
-        <ProposalHeader
-          proposal={proposal}
-          networkParams={networkParams}
-          voteState={voteState}
-        />
+        <ProposalHeader proposal={proposal} voteState={voteState} />
         <ProposalsListItemDetails proposal={proposal} />
       </RoundedWrapper>
     </li>
