+++ apps/governance/src/routes/proposals/components/vote-details/index.tsx
@@ -1,1 +1,1 @@
-export { VoteDetails } from './vote-details';
+export { UserVote } from './user-vote';
