+++ libs/ui-toolkit/src/components/icon/blueprint-icons/icon.tsx
@@ -7,7 +7,7 @@ export type { IconName } from '@blueprintjs/icons';
 export interface IconProps {
   name: IconName;
   className?: string;
-  size?: 2 | 3 | 4 | 6 | 8 | 10 | 12 | 14 | 16;
+  size?: 2 | 3 | 4 | 5 | 6 | 8 | 10 | 12 | 14 | 16;
   ariaLabel?: string;
 }
 
@@ -23,6 +23,7 @@ export const Icon = ({ size = 4, name, className, ariaLabel }: IconProps) => {
       'w-2 h-2': size === 2,
       'w-3 h-3': size === 3,
       'w-4 h-4': size === 4,
+      'w-5 h-5': size === 5,
       'w-6 h-6': size === 6,
       'w-8 h-8': size === 8,
       'w-10 h-10': size === 10,
